fixed layout and wording (issues #20 & #27)
Honestly, not certain that the aesthetics are good enough. But better
than before. (I removed <strong> from listselect, and used it only for
the headers in the labels. I also included <br> tags in the labels. In
the future a header and subhead with proper css tags in the list select
would be preferred.