Standardize interface centering
This moves the interface L/R margin definition to a mixin.

It's not added directly to the wrapping .stage__interface since the
search interface needs special handling (at least for now).

This changes the layout of the sociogram: adds back a left margin (which
was present in earlier version) and a small right margin, which prevents
ndoe labels from being cut off on the sides. Fixes #659.

Additionally, prompt margins are added to the existing mixin.