Fix NodeForm submitting
Overlay is now a HOC, so the ref needs to be restored.

See discussion at https://github.com/reduxjs/react-redux/issues/914,
which is a related issue; this is the scenario where forwardRef is described
as not possible from user code. To work around, this manually stores & gets
the Overlay instance.

Note: I wasn't able to assert the needed state with enzyme; hence the
test renderer.