DyadCensus should not defer to onComplete
When you hijack the navigation with the beforeNext hooks, ProtocolScreen caches the direction that was clicked for the next time onComplete is called.

e.g. click back (gets cached as backwards), clickin through pairs never involves the timeline, so it never gets reset to ‘forwards’, so on the last step when you call onComplete, it goes back a prompt…

This change means that DyadCensus doesn't defer to onComplete, ProtocolScreen may also need updating.