Update to Electron 18
Further preparation for fresco project. Updated to Electron 18, and
began work on restructuring the project. Added an `electron` folder,
that will contain the main and preload scripts. Removed the integration
tests, since they haven't worked for over a year now.

App is currently able to boot
inside electron when targeting web. Further refactoring work is needed
to remove use of `require`. This work will form the basis of the new
bridge API.