Use realtime canvas everywhere for better performance
This is a WIP PR which aims to unify and replace two existing systems
for positioning nodes on sociogram-like screens (Sociogram and
Narrative interfaces).

- RealtimeCanvas - renders outside of React, and was developed for the
  force directed layout feature. Very high performance. Not used on the
  Narrative Interface, or when automatic layout is disabled
- Canvas - renders inside of React, and uses the (poorly performing)
  legacy drag and drop system.

General approach is to create a LayoutContext, which provides any child
component (NodeLayout, EdgeLayout, etc) with a means to get the current
position of all nodes outside of the React render cycle.

These positions can either be the result of a force simulation, or user
provided positions (or both, in the case of a user dragging a node
within the simulation).

Left to finish:

- [ ] Dragging nodes is offset for some reason
- [ ] Edges are not displaying
- [ ] Convex hull implementation needs to be updated to use the new
  layout context, and requestAnimationFrame
- [ ] ScreenManager should be replaced with a resizeObserver (see
  useSize hook)
- [ ] Need to implement a replacement for unpositioned nodes
  (NodeBucket)
- [ ] Force simulation mode is not implemented
- [ ] Narrative interface needs to be updated to use new system