Attempt at new sortOrder
This is a first pass at a new sortOrder that includes functions that
are needed by the two mode sociogram feature. These are:

- Ability to sort by node type
- Ability to run across multiple node types, and handle the possibility
  that some nodes will not have a value for a given attribute

Node type sorting was implemented using an additional property for each
sort rule (hierachy) that should only be present on rules targetting the
`_nodeType` attribute. This still needs to be added to the protocol
schema if we decide it is the way to go.

The API of the function was changed somewhat to simplify its use. This
makes it not directly compatible with existing code, but only minor
updates should be needed.

Work has stalled because of difficulties handling undefined and null
values.

The previous version of this was built on lodash.orderBy, which puts
undefined and null values at the start of the list. This is not
what we want.

Fixing this is very difficult because:

1. orderBy does not support a custom comparator function in any
  released version. This functionality was added four years ago, but
  has not seen the light of day.
2. Without a custom comparator function, we cannot use the attribute
  type to sort null and undefined values to the end of the list (for
  example by returning -Infinity/Infinity for a number, and true/false)
  for a boolean.