Implement processProtocolSortRule
Implemented a function that provides a compatibility layer between the
old and new sort systems, and added significant additional test
coverage.

Differences:
- The old system provided a configurable "attribute path" that defaulted
to `entityAttributesProperty`. The new system requires the path to be
specified in full. This function appends the entity attributes property
to the path, but only where the path is not 'type' (used for node type
rules)
- The old system also did not account for variable type in sort
behaviour. The new system uses a 'type' property to do this. This
function maps the variable type to the new type by looking up the
variable definition in the codebook. If there is no variable definition
the rule will not be modified.