Implement experimental workaround for Android Server pairing
The new android platform introduced in 6.5.0 made the app be served from
the HTTPS context by default. This broke pairing, because requests to
Server are sent via insecure HTTP (we use our own TLS certificate to
secure the communication once pairing is complete) and chrome disallows
"mixed content" requests.

I implemented a preference here that forces the app to load from HTTP,
which "fixes" the issue, but it has some drawbacks:

- Some browser features are disabled when loaded from an insecure
  context. We don't know what these are, exactly.
- iOS uses its own preference for 'scheme', so pairing is presumably
  still broken there.